Updated: NumberUtility::wrapValue() now uses templates to respect the the type. Updated: Image updated to reflect changes in wrapValue.

+++ Image.cpp
@@ -284,7 +284,7 @@ PENJIN_ERRORS Image::assignClipAreas(CRuint xTiles,CRuint yTiles,CRuint skipTile
             //  If not in a tilesheet we can do rotation caching
             if(!sheetMode && scale.x == 1 && scale.y == 1)
             {
-                uint myAngle = NumberUtility::wrapValue(angle,359);
+                uint myAngle = NumberUtility::wrapValue((uint)angle,(uint)359);
                 if(myAngle == 0)
                     SDL_BlitSurface(images.at(i), &src, dstimg, &dst);
                 else
